♻️ create 'api' folder in lib for backend endpoints
